Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -100,7 +100,7 @@ def get_average_age_for_season(data, season)
       end
     end
 
-  return (averageAge / count + 1).to_i
+  return (averageAge / count + 1).to_f
 end
